requirements.txt
fastapi==0.109.0
uvicorn==0.27.1
psycopg2-binary==2.9.9

python-dotenv==1.0.1

pytest==9.0.1
pytest-cov==7.0.0
httpx==0.27.0
